backend/foodgram/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djoser==2.2.0
Pillow==10.0.0
django-filter==23.2
psycopg2-binary==2.9.3
gunicorn==20.1.0